config/requirement.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
MetaTrader5
python-dotenv
requests
django-cors-headers
psycopg2-binary
pandas
numpy
pytz
gunicorn
